Add support for indexed strings
This allows us to do an XPath query to add the static strings to a hash table.

By converting the strings to indexed strings we can do integer comparisons,
rather than calling strcmp() a few million times. It also allows us to skip
parts of union queries where the text() or attr() is known not to exist.

Depending on the document structure, this can speed up queries by 5 to 90%!