Don't hardcode the path of python3
Most *BSD systems don't consider python as an integral part of the
operating system, so python is seldom installed in /usr prefix.

To avoid hardcoding the path in the script itself, we ask meson to find
the location of python3 for us and use the result to run the script.
This also removes the need of shebang and the executable bit.