Convert the contents after decoding a GZipped file
Previously the contents were assumed to be XML, and this didn't work with GZip
compressed YAML.

The code isn't optimal as it writes decodes the GZipped file to a temporary
file. This was done because the converter API uses a GFile instead of a
GInputStream. The existing XML case doesn't use the temporary file so shouldn't
have a performance hit.

Fixes https://github.com/hughsie/libxmlb/issues/13