xb-machine: Don’t allocate opcodes individually
On a simple test of loading gnome-software and switching between a
couple of categories, this reduces calls to `g_slice_new0()` from around
21 million to around 5 million.

It reduces the time to show (refine) the ‘Games’ category from around 9s
to around 5s the first time it’s shown (in a very unrigorous test).

This makes `XbOpcode` stack-only, and deprecates all API which requires
it to be heap-allocated. For the most part, this only requires
mechanical changes in the rest of the codebase to operate on
stack-allocated opcodes, removing one layer of indirection in some
places.

The largest other change needed was to rewrite the optimiser to avoid
using `GArray`, as tracking the ownership of `XbOpcode`s through several
different `GArray`s was getting too complex. The new implementation of
the optimiser works as before, except it only uses one helper stack
rather than a helper stack and two temporary arrays. This brings its own
inherent performance benefits, as around 5 million XPath queries are
done when gnome-software starts up, so making the optimiser run quickly
is important.

This doesn’t introduce any ABI breaks in the library’s ABI, but does
make a few functions in the public API abort when called. Neither
gnome-software nor fwupd seem to use these functions, though, so this
doesn’t break third party software.

Signed-off-by: Philip Withnall <withnall@endlessm.com>