xb-builder-source-ctx: Store a GFile for the source too
As well as an input stream for reading the source, store a `GFile`
pointing to the on-disk file which backs the source, if possible.

This will allow the `XbBuilderSourceCtx` to load the source data
directly from disk, which may be more efficient than streaming it in.
For example, by memory mapping.

This breaks some internal API.

Signed-off-by: Philip Withnall <withnall@endlessm.com>