xb-builder-node: Only allocate attrs array when it’s needed
Before this commit, 10.2MB of `GPtrArray`s were being allocated to store
`attrs` in `XbBuilderNode`. With this commit, that’s down to 1.9MB — so
8.3MB of new `GPtrArray`s were empty and never used.

This reduces peak memory consumption in gnome-software when it’s
rebuilding xmlb caches.

From some profiling, of 500080 `XbBuilderNode`s created when rebuilding
the xmlb caches on Fedora 32 (triggered using `find
~/.cache/gnome-software -name '*.xmlb' -delete`):
 * 397949 (79%) have 0 attributes
 * 62924 (12%) have 1 attribute
 * 15523 (3%) have 2 attributes
 * 23579 (4%) have 3 attributes
 * 105 (<1%) have 4 attributes
 * 0 have ≥5 attributes

Signed-off-by: Philip Withnall <withnall@endlessm.com>