xb-builder-node: Only allocate children array when it’s needed
Before this commit, 10.2MB of `GPtrArray`s were being allocated to store
`children` in `XbBuilderNode`. With this commit, that’s down to 1.1MB — so
9.1MB of new `GPtrArray`s were empty and never used.

This reduces peak memory consumption in gnome-software when it’s
rebuilding xmlb caches.

From some profiling, of 500080 `XbBuilderNode`s created when rebuilding
the xmlb caches on Fedora 32 (triggered using `find
~/.cache/gnome-software -name '*.xmlb' -delete`):
 * 435513 (87%) have 0 children
 * 22308 (4%) have 1 child
 * 13315 (2%) have 2 children
 * 8143 (1%) have 3 children
 * 20798 (4%) have 4–1000 children
 * 3 (<1%) have 1001–2237 children
 * 0 have ≥2238 children

Signed-off-by: Philip Withnall <withnall@endlessm.com>