xb-builder-source: Add XB_BUILDER_SOURCE_FLAG_WATCH_DIRECTORY flag
This allows callers to specify that a source comes from a directory full
of other sources, and hence it would be better to monitor the directory
for changes rather than every file within it.

This reduces the number of file monitors used by gnome-software by an
order of magnitude as only the appstream and appdata directories need to
be monitored, rather than O(N) apps within them.

Signed-off-by: Philip Withnall <withnall@endlessm.com>