xb-silo: Conditionally remove XbNode cache
This cache has existed since libxmlb was originally written, and
contained a strong reference to every `XbNode` ever constructed by the
`XbSilo`, presumably to avoid the cost of constructing new `XbNode`
objects repeatedly.

Some profiling of it shows that, for the two `XbSilo`s in use by
gnome-software:
 • Silo 1: 4512 entries, hit rate 2352/6864 (34%)
 • Silo 2: 2613 entries, hit rate 1601/4214 (37%)

(After applying the previous commits in this MR, which reduce the size
and usage of the cache by about 1/3 on its previous levels.)

Most of those hits are fairly localised in time, and once a particular
piece of XML has been processed (for example, by gnome-software’s
appstream plugin), its nodes are not touched again, but continue to sit
in the cache forever.

Dropping the cache reduces locking overhead for accesses to the
`XbSilo`, and drops 1.8MB of heap memory usage from gnome-software.
(This was split between 990KB for the `XbNode` instances and 830KB for
their entries in the cache hash table.) It does not affect the speed of
jobs in gnome-software’s appstream plugin on startup, which remain the
same (to the nearest 100ms).

Unfortunately, fwupd relies on the cache memoising every `XbNode` so
that `xb_node_{get,set}_data()` can be used to side-load data into the
XML tree so it can be passed around between otherwise-disconnected parts
of the fwupd code. There’s no easy way to rearchitecture fwupd to avoid
this dual use of the `XbNode` tree as a representation of the XML and as
an implementation of the higher level objects.

To preserve compatibility, a new `XbSilo:enable-node-cache` property has
been created, enabled by default. Most clients will want to disable it
shortly after constructing an `XbSilo`.
`XB_BUILDER_COMPILE_FLAG_NO_NODE_CACHE` does this when compiling an
`XbBuilder`.

Signed-off-by: Philip Withnall <withnall@endlessm.com>