Add GError to xb_stack_pop() and xb_stack_push()
Many places were calling these and returning FALSE without setting the GError,
leading to critical warnings and crashes. The error code was also different
depending on the caller.

Set the optional GError in the low-level function for afl sanity.

Additionally, do not use g_assert_not_reached() when comparisons are not
handled as it's better to return with failure than just crash the caller.