xb-query: Use XbQueryContext/XbValueBindings in query functions
The `XbValueBindings` instance is used to provide the values for query
opcodes of type `XB_OPCODE_KIND_BOUND_*`, which will allow the `XbQuery`
itself to be reused over time and between threads. It also provides the
results limit and query flags.

The old versions of the APIs, which don’t accept a context, have been
kept for convenience.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>