xb-query: Don’t error when creating a query if the element doesn’t exist
Previously, `xb_query_new()` returned `G_IO_ERROR_INVALID_ARGUMENT` if
an element in one of its sections didn’t exist in the silo (i.e. if it
didn’t exist at all in any paths in the original XML file).

This isn’t really an error in the caller’s XPath syntax, so returning it
as an error from `xb_query_new()` is confusing. It also means that the
caller has to handle the case of “the element I’m searching for wasn’t
found” twice: once when constructing the query, and once when running
it.

Stop returning that error, and instead let it be returned when the query
is run. This means that (unless there’s another part of query
construction which depends on what’s in the silo) callers can now ignore
errors from `xb_query_new()` if they know their XPath is valid.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>