xb-query: Move silo member to a temporary parsing context
The silo member of `XbQuery` was never actually accessed after parsing
the XPath string, so move it to a temporary parsing context which can be
dropped once parsing’s complete.

This removes any concerns about it becoming a dangling pointer if the
`XbSilo` is finalised before the `XbQuery`, since the query no longer
holds a strong reference.

It also makes the query object a tiny bit smaller.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>