xb-silo: Don’t use GRWLockReaderLocker yet
It was only introduced in GLib 2.62, which isn’t available on (for
example) Debian Stable. Use the `g_rw_lock_*()` methods directly
instead. There’s no point in using `GRWLockReaderLocker` if the GLib
version is high enough, as the amount of macro boilerplate needed would
be relatively high.

This fixes building gnome-software (transitively) on Debian Stable.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>