Support adding tokens to the strtab at compile time
This allows us to normalize and tokenize at silo build time, which allows us
to perform `~=` queries on UTF-8 strings without using g_str_match_string()
in the fast path.

This allows us to search UTF-8 strings at the same speed as ASCII ones, at
the expense of a 15% larger xmlb blob.

Also add the --tokenize argument to xb-tool to allow building the silo in this
way.

    xb-tool compile test.xmlb fedora.xml.gz -v --tokenize name --tokenize summary
    xb-tool query test.xmlb "components/component/pkgname[text()='gimp']/../id"
    xb-tool query test.xmlb "components/component/name[text()~='manip']/../id"
    xb-tool query test.xmlb "components/component/name[text()~='изображ']/../id"