xb-silo: Modernise property declaration
This allows all the properties to be installed at once in the class
struct, which is a little faster.

It also allows the `GParamSpec`s to be looked up after class
initialisation, which will be used in upcoming commits.

Finally, it means that `-Wswitch-enum` will warn us if we add a new
property but forget to implement it in `get_property()` or
`set_property()`.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>