xb-silo: Clarify GMainContext usage and signal emission
An `XbSilo` can be created in one thread and queried from another. With
any object like this, it’s necessary to clarify which thread signal
emissions will be done in, otherwise any code which connects to a signal
needs to be able to run in any thread, which can easily cause all sorts
of reentrancy and deadlock problems.

So, store the thread-default `GMainContext` at the time of creating the
`XbSilo`, and ensure all signals are emitted in that thread.

This applies also to `GFileMonitor`, whose callback will be emitted in
the thread-default main context at the time of setting up the monitor.
Change creation of the monitor so that it happens in the silo’s main
context.

This helps fix issues related to a failure of silo monitoring in
gnome-software, where silos were being created in a worker thread with
an emphemeral `GMainContext`. Once the worker thread task had been
completed, file monitoring stopped working for the silo.

The fix for gnome-software
(https://gitlab.gnome.org/GNOME/gnome-software/-/merge_requests/989)
does not require this fix to libxmlb, but this fix will help avoid
similar problems in future.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>