meson: use the correct lookup method for the python3 script interpreter
The python module's find_installation method is incorrect -- it does a
variety of things geared towards building and installing python modules
into site-packages. This is slow and inefficient and completely
pointless -- all that is needed here is to run scripts.

find_program() is guaranteed to find the python3 interpreter too, and is
the correct way to find *any* other program, so use it for python as
well.