xb-builder: Release source file handles early
The XbBuilderSource::istream can be a file input stream holding a file handle
to the underlying file. These can be released early, as soon as they are
processed, thus they do not consume opened file handles of the process needed
for other operations. This is achieved by not caching the input stream inside
the XbBuilderSource structure.

This had been reported downstream:
https://bugzilla.redhat.com/show_bug.cgi?id=2244575