Fix writing files on Windows
To do this, remove the 'cached' silo from the builder -- which allows us to
drop the mmap to the existing silo before writing the new one.

This works around a regression in GLib, which started calling `_chsize()` during
`g_file_set_contents_full()` -- which doesn't work when the thing we're trying
to truncate is `mmap`'d. It's probably dumb to *not* unmap the previous file
before writing the new one -- although unmapping and then re-mapping the file
was the first thing libxmlb did after writing the file.

That's 6 hours of my life I'm never going to get back, although it does make the
builder slightly simpler.